+++ home-helper-front/src/components/atoms/Button/index.tsx
@@ -1,0 +1,1 @@
+export { Button } from "@components/atoms/Button/Button";
