+++ home-helper-front/src/components/atoms/Counter/index.tsx
