+++ home-helper-front/src/components/atoms/Layout/HouseHoldLayout.tsx
@@ -1,6 +1,6 @@
 import { FC, ReactElement } from "react";
 import { index as AppShell } from "./";
-import { LinkList } from "@components/ui/Card";
+import { LinkList } from "@components/atoms/Card";
 
 type HouseHoldLayoutProps = {
   children: ReactElement;
