+++ home-helper-front/src/components/page/charge_point/ChargePointPage.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { ChargePointContainer } from "@components/presenter/charge_point/ChargePointContainer";
-import { HomeHelperLayout } from "@components/ui/Layout/HomeHelperLayout";
+import { HomeHelperLayout } from "@components/atoms/Layout/HomeHelperLayout";
 
 export const Page: FC = () => (
   <HomeHelperLayout>
