+++ home-helper-front/src/components/page/exchange_item/ExchangeItemRegisterPage.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { HomeHelperLayout } from "@components/ui/Layout/HomeHelperLayout";
+import { HomeHelperLayout } from "@components/atoms/Layout/HomeHelperLayout";
 import { ExchangeItemRegisterContainer } from "@components/presenter/exchange_item/register/ExchangeItemRegisterContainer";
 
 export const Page: FC = () => (
