+++ home-helper-front/src/components/page/help_item/HelpItemRegisterPage.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { HomeHelperLayout } from "@components/ui/Layout/HomeHelperLayout";
+import { HomeHelperLayout } from "@components/atoms/Layout/HomeHelperLayout";
 import { HelpItemRegisterContainer } from "@components/presenter/help_item/register/HelpItemRegisterContainer";
 
 export const Page: FC = () => (
