+++ home-helper-front/src/components/page/home_helper/HomeHelperPage.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { HomeHelperContainer } from "@components/presenter/home_helper/HomeHelperContainer";
-import { HomeHelperLayout } from "@components/ui/Layout/HomeHelperLayout";
+import { HomeHelperLayout } from "@components/atoms/Layout/HomeHelperLayout";
 
 export const Page: FC = () => (
   <HomeHelperLayout navHidden>
