+++ home-helper-front/src/components/page/house_hold/HouseHoldPage.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { HouseHoldContainer } from "@components/presenter/house_hold/HouseHoldContainer";
-import { HouseHoldLayout } from "@components/ui/Layout/HouseHoldLayout";
+import { HouseHoldLayout } from "@components/atoms/Layout/HouseHoldLayout";
 
 export const Page: FC = () => (
   <HouseHoldLayout navHidden>
