+++ home-helper-front/src/components/page/point_history/PointHistoryPage.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { PointHistoryContainer } from "@components/presenter/point_history/PointHistoryContainer";
-import { HomeHelperLayout } from "@components/ui/Layout/HomeHelperLayout";
+import { HomeHelperLayout } from "@components/atoms/Layout/HomeHelperLayout";
 
 export const Page: FC = () => (
   <HomeHelperLayout>
