+++ home-helper-front/src/components/page/prosession_point/PossessionPointPage.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { PossessionPointContainer } from "@components/presenter/possession_point/PossessionPointContainer";
-import { HomeHelperLayout } from "@components/ui/Layout/HomeHelperLayout";
+import { HomeHelperLayout } from "@components/atoms/Layout/HomeHelperLayout";
 
 export const Page: FC = () => (
   <HomeHelperLayout>
