+++ home-helper-front/src/components/page/top/TopPage.tsx
@@ -1,6 +1,6 @@
 import { FC } from "react";
 import { TopContainer } from "@components/presenter/top/TopContainer";
-import { TopLayout } from "@components/ui/Layout/TopLayout";
+import { TopLayout } from "@components/atoms/Layout/TopLayout";
 
 export const Page: FC = () => (
   <TopLayout>
