+++ home-helper-front/src/components/presenter/charge_point/ChargePointPresenter.tsx
@@ -1,8 +1,8 @@
 import { FC } from "react";
 import { Flex } from "@mantine/core";
 import styles from "./styles.module.scss";
-import { UpdateNotification } from "@components/ui/UpdateNotification";
-import { Table, TbodyProps } from "@components/ui/Table";
+import { UpdateNotification } from "@components/atoms/UpdateNotification";
+import { Table, TbodyProps } from "@components/atoms/Table";
 
 type ChargePointPresenterProps = {
   fromDate: string;
