+++ home-helper-front/src/components/presenter/point_history/PointHistoryPresenter.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { Tab } from "@components/ui/Tab";
+import { Tab } from "@components/atoms/Tab";
 import { ChargedPointHistoryContainer as ChargedPointHistory } from "@components/presenter/charged_point_history/ChargedPointHistoryContainer";
 import { UsedPointHistoryContainer as UsedPointHistory } from "@components/presenter/used_point_history/UsedPointHistoryContainer";
