--- home-helper-front/src/components/ui/Button/index.tsx
@@ -1,1 +1,0 @@
-export { Button } from "@components/ui/Button/Button";
